+++ test/core/store/fs/impl/test_rasterio.py
@@ -13,12 +13,19 @@ import s3fs
 import xarray
 import xarray as xr
 
-from xcube.core.store.fs.impl.dataset import DatasetGeoTiffFsDataAccessor
-from xcube.core.store.fs.impl.geotiff import (
-    GeoTIFFMultiLevelDataset,
-    MultiLevelDatasetGeoTiffFsDataAccessor,
-)
+from xcube.core.store.fs.impl.rasterio import DatasetGeoTiffFsDataAccessor
+from xcube.core.store.fs.impl.rasterio import DatasetJp2FsDataAccessor
+from xcube.core.store.fs.impl.rasterio import RasterioMultiLevelDataset
+from xcube.core.store.fs.impl.rasterio import MultiLevelDatasetGeoTiffFsDataAccessor
+from xcube.core.store.fs.impl.rasterio import MultiLevelDatasetJp2FsDataAccessor
+
+
 from xcube.util.jsonschema import JsonSchema
+
+
+_COG_TEST_FILE = "sample-cog.tif"
+_JP2_TEST_FILE = "sample.jp2"
+_GEOTIFF_TEST_FILE = "sample-geotiff.tif"
 
 
 class RioXarrayTest(unittest.TestCase):
@@ -38,7 +45,7 @@ class RioXarrayTest(unittest.TestCase):
             "examples",
             "serve",
             "demo",
-            "sample-cog.tif",
+            _COG_TEST_FILE,
         )
         array = rioxarray.open_rasterio(cog_path)
         self.assertIsInstance(array, xr.DataArray)
@@ -69,9 +76,9 @@ class RioXarrayTest(unittest.TestCase):
         self.assertTrue(f"{e.exception}".startswith("Cannot open overview level 2 of "))
 
 
-class GeoTIFFMultiLevelDatasetTest(unittest.TestCase):
+class RasterIoMultiLevelDatasetTest(unittest.TestCase):
     """
-    A class to test wrapping of geotiff file into a multilevel dataset
+    A class to test wrapping of a file into a multilevel dataset
     """
 
     @classmethod
@@ -92,9 +99,9 @@ class GeoTIFFMultiLevelDatasetTest(unittest.TestCase):
         return fs, cog_path
 
     def test_local_fs(self):
-        fs, cog_path = self.get_params("sample-cog.tif")
-        ml_dataset = GeoTIFFMultiLevelDataset(fs, None, cog_path)
-        self.assertIsInstance(ml_dataset, GeoTIFFMultiLevelDataset)
+        fs, cog_path = self.get_params(_COG_TEST_FILE)
+        ml_dataset = RasterioMultiLevelDataset(fs, None, cog_path)
+        self.assertIsInstance(ml_dataset, RasterioMultiLevelDataset)
         self.assertEqual(3, ml_dataset.num_levels)
         self.assertEqual(cog_path, ml_dataset.ds_id)
         self.assertEqual([(320, 320), (640, 640), (1280, 1280)], ml_dataset.resolutions)
@@ -114,12 +121,12 @@ class MultiLevelDatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
     """
 
     def test_read_cog(self):
-        fs, cog_path = GeoTIFFMultiLevelDatasetTest.get_params("sample-cog.tif")
+        fs, cog_path = RasterIoMultiLevelDatasetTest.get_params(_COG_TEST_FILE)
         ml_data_opener = MultiLevelDatasetGeoTiffFsDataAccessor()
         ml_dataset = ml_data_opener.open_data(
             cog_path, fs=fs, root=None, tile_size=[256, 256]
         )
-        self.assertIsInstance(ml_dataset, GeoTIFFMultiLevelDataset)
+        self.assertIsInstance(ml_dataset, RasterioMultiLevelDataset)
         self.assertEqual(3, ml_dataset.num_levels)
         dataset = ml_dataset.get_dataset(0)
         self.assertEqual(["band_1", "band_2", "band_3"], list(dataset.data_vars))
@@ -132,15 +139,34 @@ class MultiLevelDatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
         self.assertIsInstance(ml_data_opener.get_open_data_params_schema(), JsonSchema)
 
     def test_read_geotiff(self):
-        fs, tiff_path = GeoTIFFMultiLevelDatasetTest.get_params("sample-geotiff.tif")
+        fs, tiff_path = RasterIoMultiLevelDatasetTest.get_params(_GEOTIFF_TEST_FILE)
         data_opener = MultiLevelDatasetGeoTiffFsDataAccessor()
 
         ml_dataset = data_opener.open_data(
             tiff_path, fs=fs, root=None, tile_size=[512, 512]
         )
-        self.assertIsInstance(ml_dataset, GeoTIFFMultiLevelDataset)
+        self.assertIsInstance(ml_dataset, RasterioMultiLevelDataset)
 
         self.assertEqual(1, ml_dataset.num_levels)
+        dataset = ml_dataset.get_dataset(0)
+        self.assertEqual((1387, 1491), dataset.band_1.shape)
+
+
+class MultiLevelDatasetJp2FsDataAccessorTest(unittest.TestCase):
+    """
+    A class to test JPEG 2000 for multilevel dataset opener
+    """
+
+    def test_read(self):
+        fs, file_path = RasterIoMultiLevelDatasetTest.get_params(_JP2_TEST_FILE)
+        data_opener = MultiLevelDatasetJp2FsDataAccessor()
+
+        ml_dataset = data_opener.open_data(
+            file_path, fs=fs, root=None, tile_size=[512, 512]
+        )
+        self.assertIsInstance(ml_dataset, RasterioMultiLevelDataset)
+
+        self.assertEqual(3, ml_dataset.num_levels)
         dataset = ml_dataset.get_dataset(0)
         self.assertEqual((1387, 1491), dataset.band_1.shape)
 
@@ -152,7 +178,7 @@ class DatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
     """
 
     def test_ml_to_dataset(self):
-        fs, cog_path = GeoTIFFMultiLevelDatasetTest.get_params("sample-cog.tif")
+        fs, cog_path = RasterIoMultiLevelDatasetTest.get_params(_COG_TEST_FILE)
         data_accessor = DatasetGeoTiffFsDataAccessor()
         self.assertIsInstance(data_accessor, DatasetGeoTiffFsDataAccessor)
         self.assertEqual("geotiff", data_accessor.get_format_id())
@@ -165,7 +191,7 @@ class DatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
         )
 
     def test_dataset(self):
-        fs, tiff_path = GeoTIFFMultiLevelDatasetTest.get_params("sample-geotiff.tif")
+        fs, tiff_path = RasterIoMultiLevelDatasetTest.get_params(_GEOTIFF_TEST_FILE)
         data_accessor = DatasetGeoTiffFsDataAccessor()
         self.assertIsInstance(data_accessor, DatasetGeoTiffFsDataAccessor)
         self.assertEqual("geotiff", data_accessor.get_format_id())
@@ -179,23 +205,75 @@ class DatasetGeoTiffFsDataAccessorTest(unittest.TestCase):
         self.assertIsInstance(dataset, xarray.Dataset)
 
 
-class ObjectStorageMultiLevelDatasetTest(S3Test):
+class DatasetJp2FsDataAccessorTest(unittest.TestCase):
     """
-    A Test class to test opening a cog file from AWS S3
+    A Test class to test opening a JPEG 2000 as multilevel dataset or
+    as normal dataset
     """
 
-    def test_s3_fs(self):
+    def test_ml_to_dataset(self):
+        fs, file_path = RasterIoMultiLevelDatasetTest.get_params(_JP2_TEST_FILE)
+        data_accessor = DatasetJp2FsDataAccessor()
+        self.assertIsInstance(data_accessor, DatasetJp2FsDataAccessor)
+        self.assertEqual("jpeg2000", data_accessor.get_format_id())
+        dataset = data_accessor.open_data(
+            data_id=file_path, overview_level=1, fs=fs, root=None, tile_size=[512, 512]
+        )
+        self.assertIsInstance(dataset, xarray.Dataset)
+        self.assertIsInstance(
+            data_accessor.get_open_data_params_schema(file_path), JsonSchema
+        )
+
+    def test_dataset(self):
+        fs, file_path = RasterIoMultiLevelDatasetTest.get_params(_JP2_TEST_FILE)
+        data_accessor = DatasetJp2FsDataAccessor()
+        self.assertIsInstance(data_accessor, DatasetJp2FsDataAccessor)
+        self.assertEqual("jpeg2000", data_accessor.get_format_id())
+        dataset = data_accessor.open_data(
+            data_id=file_path,
+            fs=fs,
+            root=None,
+            tile_size=[256, 256],
+            overview_level=None,
+        )
+        self.assertIsInstance(dataset, xarray.Dataset)
+
+
+class ObjectStorageMultiLevelDatasetTest(S3Test):
+    """
+    A Test class to test opening files with rasterio from AWS S3
+    """
+
+    def setUp(self) -> None:
         region_name = "eu-central-1"
-        s3 = s3fs.S3FileSystem(region_name=region_name, anon=True)
-        data_id = "xcube-examples/sample-cog.tif"
-        ml_dataset = GeoTIFFMultiLevelDataset(s3, None, data_id)
+        self._s3 = s3fs.S3FileSystem(region_name=region_name, anon=True)
+
+    def test_s3_fs_tif(self):
+        data_id = f"xcube-examples/{_COG_TEST_FILE}"
+        ml_dataset = RasterioMultiLevelDataset(self._s3, None, data_id)
         self.assertEqual(3, ml_dataset.num_levels)
         self.assertEqual((343, 343), ml_dataset.get_dataset(0).band_1.shape)
-        dataset_path = "xcube-examples/sample-geotiff.tif"
+        dataset_path = f"xcube-examples/{_GEOTIFF_TEST_FILE}"
         data_accessor = DatasetGeoTiffFsDataAccessor()
         dataset = data_accessor.open_data(
             data_id=dataset_path,
-            fs=s3,
+            fs=self._s3,
+            root=None,
+            tile_size=[256, 256],
+            overview_level=None,
+        )
+        self.assertIsInstance(dataset, xarray.Dataset)
+
+    def test_s3_fs_jp2(self):
+        data_id = f"xcube-examples/{_JP2_TEST_FILE}"
+        ml_dataset = RasterioMultiLevelDataset(self._s3, None, data_id)
+        self.assertEqual(3, ml_dataset.num_levels)
+        self.assertEqual((1387, 1491), ml_dataset.get_dataset(0).band_1.shape)
+        dataset_path = f"xcube-examples/{_JP2_TEST_FILE}"
+        data_accessor = DatasetJp2FsDataAccessor()
+        dataset = data_accessor.open_data(
+            data_id=dataset_path,
+            fs=self._s3,
             root=None,
             tile_size=[256, 256],
             overview_level=None,