+++ xcube/plugin.py
@@ -106,6 +106,12 @@ _FS_DATA_ACCESSOR_ITEMS = (
     ),
     ("geodataframe", "shapefile", "gpd.GeoDataFrame in ESRI Shapefile format"),
     ("geodataframe", "geojson", "gpd.GeoDataFrame in GeoJSON format"),
+    ("dataset", "jpeg2000", "xarray.Dataset in JPEG 2000 format"),
+    (
+        "mldataset",
+        "jpeg2000",
+        "xcube.core.mldataset.MultiLevelDataset in JPEG 2000 format",
+    ),
 )
 
 _FS_DATA_OPENER_ITEMS = _FS_DATA_ACCESSOR_ITEMS