+++ CHANGES.md
@@ -1,4 +1,11 @@
-## Changes in 1.11.2 (in development)
+## Changes in 1.12.0 (in development)
+
+### Enhancements
+
+* support reading JPEG2000 files as datasets or multi-level datasets from file
+  systems. Openers and writers are available as `dataset:jpeg2000:<protocol>` or
+  `mldataset:jpeg2000:<protocol>`, where protocol is the respective file system. 
+  These openers and writers are available to any file system data store.
 
 ### Other changes